fix(analysis): 修改AvgIntervalMs字段类型为float64
将KeyStat结构体中的AvgIntervalMs字段从int类型改为float64类型，
以支持更精确的时间间隔计算和存储。

+++ internal/service/analysis/analysis.go
@@ -29,7 +29,7 @@ type KeyStat struct {
 	TotalHits     int     `json:"total_hits"`
 	TotalErrors   int     `json:"total_errors"`
 	ErrorRate     float64 `json:"error_rate"` // computed
-	AvgIntervalMs int     `json:"avg_interval_ms"`
+	AvgIntervalMs float64 `json:"avg_interval_ms"`
 }
 
 // Summary represents an overall practice summary for a user.
